+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ v1.8.3
 
 *Release date: In development*
 
+- Migrate prepare_param logic from Toolium-Telefonica
+
 v1.8.2
 ------